refactor(handlers): extract write-helpers.ts from intent.ts (Stage B)
Move content-type negotiation, DDIC metadata write props, buildCreateXml,
the pre-write lint/syntax/RAP-preflight gates, package enforcement, and the
server-driven-object write engine (919 lines, 2 clusters) into one module
shared by SAPWrite + SAPLint. buildCreateXml + stripFmParamCommentBlock
re-exported for back-compat. Drops a stale unused test import. 3672 tests,
A1 zero-diff.
